+++ client/src/App.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Main from './pages/Main';
 import Share from './pages/Share';
+import Posting from './pages/Posting';
 
 const Register = React.lazy(() => import('./pages/Register'));
 const Admin = React.lazy(() => import('./pages/admin/Admin'));
@@ -14,6 +15,7 @@ function App() {
           <Route exact path="/register" component={Register} />
           <Route exact path="/admin" component={Admin} />
           <Route exact path="/share" component={Share} />
+          <Route exact path="/posting" component={Posting} />
         </React.Suspense>
       </Switch>
     </Router>
