+++ client/src/index.tsx
@@ -1,16 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
 import App from './App';
+import store from './state/store';
 import reportWebVitals from './reportWebVitals';
 import 'gestalt/dist/gestalt.css';
 import './index.css';
-
-const rootReducer = (state, action) => {
-  return state;
-};
-const store = createStore(rootReducer);
 
 ReactDOM.render(
   <React.StrictMode>
