+++ client/src/pages/Posting.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Posting = () => {
+  return (
+    <div>
+      <input type="file" />
+      <input type="text" />
+      <textarea name=""></textarea>
+    </div>
+  );
+};
+
+export default Posting;
