+++ client/src/state/action-creators/index.ts
@@ -9,3 +9,17 @@ export const fetchItems = () => async (dispatch: Dispatch) => {
     payload: response.data,
   });
 };
+
+export const setItems = (items) => {
+  return {
+    type: ActionType.GET_METHOD,
+    payload: items,
+  };
+};
+
+export const selectedItem = (item) => {
+  return {
+    type: ActionType.SELECTED_ITEM,
+    payload: item,
+  };
+};
