+++ client/src/state/reducers/index.ts
@@ -6,3 +6,5 @@ const reducers = combineReducers({
 });
 
 export default reducers;
+
+export type State = ReturnType<typeof reducers>;
